+++ dsmr_datalogger/admin.py
@@ -1,16 +1,33 @@
 from django.contrib.admin.filters import DateFieldListFilter
 from django.contrib import admin
+from django.utils.translation import ugettext_lazy as _
 from solo.admin import SingletonModelAdmin
 
 from dsmr_backend.mixins import ReadOnlyAdminModel
-from .models.settings import DataloggerSettings
+from .models.settings import DataloggerSettings, RetentionSettings
 from .models.reading import DsmrReading
-from dsmr_datalogger.models.statistics import MeterStatistics
+from .models.statistics import MeterStatistics
 
 
 @admin.register(DataloggerSettings)
 class DataloggerSettingsAdmin(SingletonModelAdmin):
     list_display = ('com_port', )
+
+
+@admin.register(RetentionSettings)
+class RetentionSettingsAdmin(SingletonModelAdmin):
+    list_display = ('data_retention_in_hours', )
+    fieldsets = (
+        (
+            None, {
+                'fields': ['data_retention_in_hours'],
+                'description': _(
+                    'Detailed instructions for configuring data retention can be found here: '
+                    '<a href="https://dsmr-reader.readthedocs.io/nl/latest/retention.html">Retention documentation</a>'
+                )
+            }
+        ),
+    )
 
 
 @admin.register(DsmrReading)